♻️ Split router setup.
